Park func_002609E8 — isnan reg alloc near-miss

+++ tough_nuts/func_0023F7B0/13F7B0.c
@@ -1,12 +1,13 @@
+
 int *func_0023F7B0(int *a0, long long a1)
 {
-    unsigned int u;
-    int s;
-    u = (unsigned int)a1;
-    a0[2] = 0x53;
-    s = (int)u;
-    a0[0] = s;
-    a0[1] = 0;
-    a0[3] = 0;
-    return a0 + 4;
+  long long u;
+  int s;
+  u = (unsigned int) a1;
+  a0[2] = 0x53;
+  s = (int) u;
+  a0[0] = s;
+  a0[1] = 0;
+  a0[3] = 0;
+  return a0 + 4;
 }
